seperated detailView's components
