+++ PriceTracker/PriceTracker/Views/Detail/DetailView.swift
@@ -36,111 +36,3 @@ struct DetailView: View {
         }
     }
 }
-
-fileprivate struct BasicInfoView: View {
-    private let item: DetailModel
-    
-    init(item: DetailModel) {
-        self.item = item
-    }
-    
-    var body: some View {
-        HStack {
-            Spacer()
-            AsyncImage(url: URL(string: item.info.thumb)) { image in
-                image.resizable().frame(width: 150, height: 100)
-            } placeholder: {
-                ProgressView()
-            }
-            Spacer()
-            VStack(alignment: .leading) {
-                Text(item.info.title)
-                    .font(.title2)
-                    .lineLimit(4)
-                Text("Retail Price: $\(item.deals.first?.retailPrice ?? "")")
-                Text("Cheapest Ever: $\(item.cheapestPriceEver.price)")
-            }
-            Spacer()
-        }
-    }
-}
-
-fileprivate struct DealsListView: View {
-    private let items: [Deal]
-    
-    init(items: [Deal]) {
-        self.items = items
-    }
-    
-    var body: some View {
-        List(items, id: \.dealID) { deal in
-            if let url = URL(string: "https://www.cheapshark.com/redirect?dealID=\(deal.dealID)") {
-                NavigationLink(
-                    destination: WebViewForDetailView(url: url)
-                ) {
-                    DealsView(item: deal)
-                }
-            }
-        }
-        .listStyle(.plain)
-    }
-}
-
-fileprivate struct DealsView: View {
-    private let item: Deal
-    
-    init (item: Deal) {
-        self.item = item
-    }
-    
-    var body: some View {
-        HStack {
-            ThumbnailView(urlString: "https://www.cheapshark.com/img/stores/logos/\((Int(item.storeID) ?? 0) - 1).png")
-            DealsTextView(item: item)
-                .padding(.vertical, 5)
-        }
-        .contentShape(Rectangle())
-        .background(Color(UIColor.systemBackground))
-        .cornerRadius(10)
-    }
-}
-
-fileprivate struct ThumbnailView: View {
-    private let urlString: String
-    
-    init(urlString: String) {
-        self.urlString = urlString
-    }
-    
-    var body: some View {
-        AsyncImage(url: URL(string: urlString)) { image in
-            image
-                .resizable()
-                .frame(width: 50, height: 50)
-                .scaledToFit()
-        } placeholder: {
-            ProgressView()
-        }
-    }
-}
-
-private struct DealsTextView: View {
-    private let item: Deal
-    
-    init(item: Deal) {
-        self.item = item
-    }
-    
-    var body: some View {
-        VStack(alignment: .leading) {
-            Text("Price: $\(item.price)")
-                .font(.body)
-            Text("Savings: " +
-                 ((Double(item.savings) ?? 0.0) < 1.0
-                  ? "None"
-                  : "\(String(format: "%.2f", Double(item.savings) ?? 0.0))%"))
-            .font(.subheadline)
-            .foregroundColor(.secondary)
-        }
-    }
-}
